New interface version V2 to expose more info on abstract types
The new interface V2 contains a sub-module RAWTYPES that can be used
to extract information for abstract types.

Interfaces are versioned, V1 is the former interface, while V2 introduces a
new interface, better for long term compatibility. `open Ezcmd.V2` should
be used before, and functions are in a sub-module `EZCMD`

The type `command` is replaced by the type `sub`, abstract by default, that
should be created by `EZCMD.sub`.

Functions in V2 can have an optional `~version` argument,
corresponding to the first version in which this command/argument
appeared (for documentation).